backend/requirements.txt
djangorestframework==3.15.2
django-environ==0.12.0
drf-spectacular==0.27.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0

pillow==11.1.0
psycopg2==2.9.10
python-decouple==3.8
django-q


#Django==5.2
#pillow==11.1.0
#psycopg2==2.9.10
#python-decouple==3.8

# To install the dependencies, run the following command in your terminal:
# pip install -r requirements.txt

#run backend
#python manage.py runserver

#run frontend
#npm run dev

